requirements.txt
google-generativeai
langchain
streamlit
python-dotenv
PyPDF2
langchain-google-genai


-e .